Text_To_Speech_Covertor
The "Text-to-Speech Web Application" is a simple and interactive tool built using HTML, CSS, and JavaScript. It allows users to input text into a text area and convert it into natural-sounding speech. The application supports customizable voice options, including Indian English, and lets users adjust the pitch for a personalized listening experience. With its user-friendly interface and responsive design, this project demonstrates the practical implementation of the Web Speech API for accessibility and engagement.